serious `0.4` (`Toolips` `0.3`) steps :)
Getting close to this one. `Toolips` and `ToolipsServables` essentially done -- though I am going to do more with parallel processing in `Toolips`. and still have tests to do on that. Then tests for this, docs for this, finish this, get this working TIP TOP. Along with this `ToolipsSVG` next.
not sure if `SwipeMap` is in here, it should be !

`Gattino` -> More