Updated `document_linker`, `on_start` binding
nearly completed the basis for the `0.3` extension. Gotta say, `0.3` is clearly faster to do this with. One exciting thing is just how much the data has been cut down, one problematic part of this is I have yet to understand how I am going to handle RPC. Perhaps I will make the `Vector{Event}` in the `Session` `Dict` into a `Vector{AbstractEvent}`. While abstraction sucks and I hate to do it in a structure period, I think this will allow me to get rid of the peer table, and register the event. This will also allow for event extensions, so I think this might be ideal :)
That being said, I *have* cracked open `ToolipsSVG`, though I have not done much for it. We are getting there! `0.4` + `0.3` is on the horizon!